Fix multifile consumption build
The existing logic that was trying to output virtual slots if
`factory.CompilationModuleGroup.ShouldProduceFullType(_type)` was only
okay for multifile production builds. In consumption builds, we might
see unreferenced vtable slots anywhere in the hierarchy. The fix is to
propagate `relocsOnly` down to vtable emission.

The nice side effect of this fix is that we no longer need to report the
generic dictionary as a non-reloc dependency.