Count vtable slots declaratively (#2931)
Just a quick cleanup - `OutputVirtualSlotAndInterfaceCount` has a bunch
of nontrivial logic in it that needs to match exactly with what
`OutputVirtualSlots` will do. We can get by without this duplication by
just counting the number of slots `OutputVirtualSlots` emitted.

I also added an assert that checks `GetVTableOffset` is not bogus while
I was at it.