Fix generics with TLS fields in multi-module
The previous commit to add multi-module support for thread statics
(https://github.com/dotnet/corert/commit/be92d719) has a small issue
causing a link failure if the generic type containing the thread-static
field was instantiated in multiple modules.

That commit changed from referring to the thread static base directly to
referring to the type thread static index, which is looked up on a list
reached through the type manager. This mechanism ensures that even
though we put a copy of the static base in multiple object files, we'll
unify because the thread static index nodes are COMDAT folded. The
ThreadStaticsNodes themselves can't be folded since they're embedded
objects in a list, so change the naming to make them unique across
modules.