Make VirtualMethodUse nodes only track newslot methods (#2964)
The rest of the analysis assumes the methods are used through the
declaring slot. We make sure of that when reading inputs (e.g. a
callvirt through the non-declaring method gets normalized to the
declaring slot). We were not enforcing this which resulted in seeing
garbage slots in the vtable.