Port SecureString to CoreRT
This change ports the CoreCLR implementation of SecureString over to CoreRT. This was an attempt to create an almost verbatim copy of the CoreCLR SecureString, so that it may be moved to the Shared partition. There are a few differences in SecureString.Windows.cs that have not been resolved.

This required inclusion of a few more Win32 APIs: CryptProtectMemory, WideCharToMultiByte, ZeroMemory, SysAllocStringLen, and SysStringLen. This also required additions to PInvokeMarshal: AllocHGlobal, AllocCoTaskMem, FreeHGlobal, and FreeCoTaskMem. These have also been added to the internal Marshal class in Interop.manual.cs, which just calls the implementations in PInvokeMarshal.

[tfs-changeset: 1650789]