Optimize calls to enum's GetHashCode/Equals
Generate synthetic instance methods into which constrained calls
can resolve. This prevents boxing 'this' that would be required
before a call to the System.Enum's implementation.

This is specifically not a synthetic virtual override. The thunk
only has very little perf advantage over System.Enum's implementation
if 'this' is already boxed and is not worth the size overhead.

This is an extended port of dotnet/coreclr#7895

Fixes #2111.