Fix a race in thread shutdown on Unix (#3073)
* Fix a race in thread shutdown on Unix

The problem is that the _stopped event (that is used for signaling that
the thread has stopped) is allocated in the CreateThread() method after an
OS thread is created. As a result, the parent thread can be scheduled out
before it allocates an event and assigns it _stopped. At same time, the
newly created thread gets a chance to run, finishes its work promptly and
exits; while exiting, it tries to signal the _stopped event but it crashes
instead because the event has not been allocated yet. I believe the
problem with lifetime management of the _stopped event is due to the fact
that the original code relied on the value of _stopped in the HasStarted()
method to indicate whether a thread has been created/started or not. On
Windows, the HasStarted() method relied on the value of _osHandle for the
same purpose.

This change fixes this problem by:
1. Changing the HasStarted() method to use the explicit state of the tread
ThreadState.Unstarted (i.e. the _threadState field). This also makes
implementation of this method platform independent. The existing code
already sets the thread state to ThreadState.Unstarted (when an instance
of the RuntimeThread class is created) and clears the bit when the thread
starts executing managed code.
2. Changing RuntimeThread on Unix to allocate the stopped event before
thread creation to actually fix the race.

I have also partially enabled the BasicThreading test for Unix and added a
small regression test for the race condition. Full functionality of
BasicThreading will be enabled once we have ThreadPool supported on all
platforms.

* PR Feedback

* Fix issue with __cxa_thread_atexit in PAL. The change (https://github.com/dotnet/corert/pull/2531)
that introduced __cxa_thread_atexit in PAL did it incorrectly. The __cxa_thread_atexit function
should be called for each TLS object that needs its destructor to be executed at thread shutdown time.
The incorrect usage of __cxa_thread_atexit resulted in hangs in SuspendEE and random corruptions.