Delete ContainsGenericVariables
This is not used anywhere, and whenever I saw peoeple using it in the
past, it was for the wrong reason (people tend to use it as a more
expensive version of `IsGenericDefinition`). We don't really instantiate
things over generic parameters anywhere. Better to just delete this and
free up some flags.