Implement Assembly.GetExecutingAssembly
This replaces calls to Assembly.GetExecutingAssembly with a call to a small stub that gets the owning type of the method and from it the assembly. I'm doing this instead of just generating a hardcoded string that we `Assembly.Load` (that the `Type.GetType` transform pretty much does) because this will also do the right thing in the presence of AssemblyLoadContext.

This requires that owning types of methods that call GetExecutingAssembly need to be marked reflectable. I'll be throwing that as a bug on someone else. We might want to implement this as an intrinsic in the flow analysis as well.

[tfs-changeset: 1652072]