Add a simple thread pool implementation for Unix
This change adds a very primitive (temporary) implementation of Thread
Pool to allow Tasks to be used on Unix until a proper implementation is
available.

For simplicity of the state management, the thread pool pre-creates all
thread pool workers (controlled by the MaxThreadCount const) on the first
request and then uses a semaphore to release threads as new requests come
in.

In addition, this change enables all test cases in the BasicThreading test.