Implement MethodBase.GetCurrentMethod
Similar implementation as `Assembly.GetExecutingAssembly` - rewrite calls to `MethodBase.GetCurrentMethod` to call a helper that `ldtoken`s the calling method. Contrary to what a reasonable person would assume, `GetCallingMethod` gives you an uninstantiated method even if the method or it's containing type is generic. This was likely more to simplify implementation on the CLR than to make a sane API.

My implementation of this is kind of dumb because the runtime format that `RuntimeMethodHandle` uses cannot express generic method definitions, so in order to cover the case of `GetCallingMethod` called from a generic method without hitting UTC, MDIL and RuntimeMethodHandle format limitations (yes, the whole trifecta), the helper needs to be generic and fully capture the generic context of the calling method. David says this is 1+ week to fix and won't happen any time soon, so here we are.

The helpers on the classlib side need to uninstantiate the thing that the compiler-generated helpers loaded.

[tfs-changeset: 1652327]