NS2.0 Add back the remaining Char members, then move it to shared partition.
I realize that from CoreRT's pov, it looks like it's
inheriting even more CoreCLR aesthetic badness.

The purpose here is to minimize diff's between CoreCLR
and CoreRT so the mirror bot PR will go through smoothly.

When this gets mirrored over by the bot, the delta to CoreCLR's Char.cs
will be limited to:

- Cleaned up "using's"

- "categoryForLatin" static field gets the "s_" prefix.

- Gets CoreRT's IsSymbol() microoptimization (cache s[index] in local.)

Giving the file as a whole a bubble bath to make it presentable
is out of the scope of this commit (it'll be more efficient to
do once it's shared if anyone wants to grab that task.)