NS2.0 Add String.Intern
This won't cover literals but otherwise, gets the
job done.

Note that LockFreeReaderHashTable was declared public
(and so were a ton of classes that derived from it,
so I couldn't just make it private.)

Including it in CoreLib.dll means no one above
CoreLib can include it - they'll just the public
one from CoreLib.dll.