Globally replace "PreviousCallContainsUserCode" with "PreviousCallContainsDebuggerStepInCode" in ProjectN
To demonstrate completeness of the switch I have removed the "PreviousCallContainsUserCode" attribute altogether. The change is mechanical but it's lengthy as doing this properly required updating the various contracts and rebuilding IL transform CIBs as the well-known type set has changed. There is a bit of additional churn as apparently the contracts haven't been rebuilt for some time but thankfully not much. In McgRedirect, I have switched the generated PInvoke helper method to use the same technique I previously used for delegate Invoke methods - switching over to use the native F00F00 / FEEFEE annotations. We're currently discussing how to make a similar change in MCG code overall.

Thanks

Tomas

[tfs-changeset: 1652651]