Implement MethodBase.GetCurrentMethod
This is on the same plan as Assembly.GetExecutingAssembly (redirect
calls to this method to call a small thunk that loads what "current"
means and dispatches to the CoreLib to do the rest of the work).

The difference from `GetExecutingAssembly` is that I made the method be
a member of the type that owns the calling method so that we don't have
to come up with weird name manglings.