Add concept of "prefix mangled type/method" (#3173)
* Add concept of "prefix mangled type/method"

For type system entities (method/types) that are generated by the
compiler, we don't have a good way to guarantee a unique mangled name.
In the ECMA-based world, we rely on enumerating everything in the
assembly to generate a guaranteed non-conflicting name. On the
compiler-generated side, we don't have a good way to do such enumeration
because there's no central database of those things.

We were using multiple techniques to guarantee nonconflicting names:

1. Counter (problematic for stability when we go multithreaded)
2. Wrapping in a custom node type with special rules (extra code)
3. Getting a mangled name to use as type system name (in delegate
marshalling - kind of odd thing to do)
4. Hoping that two types names just won't conflict (I think there are
such patterns still remaining in struct marshalling)

I was in a place where I started inventing method 5 (for
`MethodBase.GetCurrentMethod` intrinsic expansion).

What we really want in all those places is just to say: "use the mangled
name of something we know we can generate a nonconflicting name for, and
append some prefix".

This change adds just that, and switches 1 and 2 from the above list to
the new plan.