Add support for RuntimeMethodHandle of a generic method definition (#3152)
This is actually two fixes:

* Support for reading/writing `ldtoken method` of a generic method
definition (only really doable from IL, so not a big deal, but I need it
elsewhere)
* Support for generating and reading a dynamic `RuntimeMethodHandle` for
a generic method definition (`MethodBase.get_MethodHandle` of a generic
method definition). We were generating a `RuntimeMethodHandle`, but it
was a handle to something weird that wasn't roudtrippable.

Both cases basically represent these as something with a null
instantiation now.