Add support for arrays of pointer types (#3192)
Progress towards dotnet/corefx#17480.

On the compiler side:
* Unblock generation of arrays of pointer EETypes. We were already doing
the right thing w.r.t. the interface lists (these *do not* implement the
generic interface: `int*[]` does not implement `IEnumerable<int*>`).
Nothing else was missing.

On the classlib side:
* I groveled through the places where we would box/unbox array element
types, or assume that if the element type is not a value type, it's a
reference type. Hopefully, I didn't miss anything.
* I pre-emptively fixed dotnet/coreclr#10646 in my implementation of
`Array.Copy`.

On the runtime side:
* I verified this doesn't affect casting (i.e. we don't allow casting
`uint*[]` to `int*[]`), so no changes should be needed.