NS2.0 Complete System.IO.Stream surface area
ObjectInvariant - an obsolete method for supporting
contracts...

CreateWaitHandle - not too useful by itself,
but I assume this was meant to be overridden
by something more useful.

Synchronized - creates SyncStream.

-

Keeping Stream structured the same as the
CoreCLR version. Not going to attempt to share
this as there still lots of scary looking
async differences. I'll leave this to the
area owner.