Use 32bit relative pointers in ExternalReferenceTable
We can't use IMAGE_REL_BASED_ADDR32NB (because it's Windows only), but
we can use IMAGE_REL_BASED_REL32 which is close enough.

For CppCodegen we'll need to stick to pointer-sized entries (that are
bigger and require an extra reloc), but CppCodegen is not far enough to
use this yet anyway.

We might also just want to just standardize on IMAGE_REL_BASED_REL32 for
any TypeManager-based modules.