Fix rules for when metadata token should be used in mapping tables
New code from TFS changed the way we refer to members from field/method
reflection mapping tables. It conservatively chose to not use tokens,
but fall back to full signatures at all times (these are normally only
needed for crossmodule references).

Fixing the logic up because this was causing failures in the rolling
build (one of the negative tests that deliberately uses bad types was
crashing the compiler because the bad type can't be used in a native
layout signature).

This fix also recovers a file size/throughput regression.

The bigger picture problem is that we need to do something about
resiliency against bad inputs at the native layout writing phase.