Reenable randomized hashing perf mitigation for CoreRT. (#3289)
This reenables the FEATURE_RANDOMIZED_STRING_HASHING paths for
Dictionary (without the #if ceremony.) The purpose of this
code is mitigate the perf regression introduced by randomized
string hashing on Dictionaries. Dictionaries only switch
to the randomized hash if they suspect a DOS attack
in progress.

String.GetLegacyNonRandomizedHashCode() is cut-paste directly
from CoreCLR (sans the ThisAssembly.DailyBuildNumber salt
which wasn't doing what it intended anyway since
ThisAssembly.DailyBuildNumber never changes in CoreCLR.)

The part that will look strange are the changes to TypeDesc.cs.
Without this, instantiating a Dictionary triggers this unfortunate
reentrancy.


 EqualityComparerHelpers.GetComparer()

   which triggers

 TryBuildGenericType

   which triggers

 another TryBuildGenericType

   which triggers

 TypeDesc.IsArray => this.GetType() == typeof(ArrayType)

   which triggers

 two Type creations

   which triggers

 RuntimeTypeInfo.EstablishDebugName

   which triggers

 ReflectionTrace.get_Enabled

   which triggers

 ReflectionEventSource class constructor

   which triggers a bunch more EventSource-related stuff
   which eventually triggers

 EqualityComparerHelpers.GetComparer() again,

   which eventually ends up apparently creating/registering
   the very same type that the first GetComparer() was trying to add.

   As a result, this all unwinds and

 RegisterDynamicGenericTypesAndMethods()

   throws an throw new ArgumentException(SR.Argument_AddingDuplicate);


I chose to break this cycle by removing the GetType()/typeof() calls
from the path and replacing them with "is" checks (after checking
that each target type is sealed. Though "is" is still probably
correct even if unsealed.)

Let me know if there's a more holistic solution for this but
in general, not triggering Type creations in low-level code
is a good practice in Project N.