Fix two CoreRT issues/
https://github.com/dotnet/corert/issues/3063

  CustomAttributeInstantiatior should stop throwing MissingMetadataExceptions


https://github.com/dotnet/corert/issues/2941

  FieldInfo.GetCustomAttributes(bool) shouldn't check if the ctor is public

Note: CoreCLR actually checks visibility relative to the decorated
member, but I'm not going to worry about that now (this is not the
right place in the code to do that anyway - the visibility check
is used to filter the enumeration.) I think we can get away without that.

While researching this, I found out that custom attribute fields
and properties *do* have to be public so I made that fix too.