This is the change to fix the GC info cliff for huge methods. For instance, app \\fxcore\apps\WindowsStore\UWP\Refresh\apps\x86\aa3825a1-2d97-4c71-8c99-578135b28cd8\00000000-0000-0000-0000-000000000000 produces a native dll that is 128 MB in size, of which 73 MB is GC info.
With this change, the size of GC info for this app reduces to about 4 MB, and the app file size to about 59 MB. Time spent in rhbind goes from 122 seconds to 11 seconds, measured on a 3.6 GHz Xeon.

Now this app is a total outlier - typical numbers for apps are a 10 - 20% reduction in GC info size (upper part of the range on x86, lower part for x64 and ARM), 0.5 - 1 % reduction in executable file size, and negligible impact on rhbind time.

Explanation of the problem:

The underlying problem causing the huge expansion in GC info is the fact that the GC roots for each call site are encoded independently. While two call sites with the exact same set of GC roots will share a GC root string, any difference will cause a separate root string to be generated. This is costly and inefficient if the two call sites share a large number of GC roots, as is often the case in large methods.

The solution is to:
   1. Expand the GC encoding to express the notion of GC root strings common to several call sites.
   2. Change the binder's GC info generation algorithms to efficiently compute sets of variables common to a set of call sites.
   3. Change the decoding logic in the runtime (and rhdump) so it's able to switch from decoding a set of GC roots specific to a call site to a common GC root set.

The implementation I have adopted for issue 1 consists of two parts:
   a) add the ability to list GC root sets common to several call sites before describing the offsets and root strings of the call sites themselves. These common GC root strings are considered implicitly numbered and hence can be referred to by an index.
   b) add the ability for a GC root string to refer to one of the common root strings by index. This is done by using a non-sensical variation of the "local slot set encoding", where the set of local slots is empty. This encoding must always be the last one in a GC root string, so it is analogous to a "tail call" if you view the GC root encoding string as instructions to the GC root listing algorithm.

For issue 2, I have a pretty clever algorithm in the binder that recursively subdivides ranges of call sites and divides the variables into three sets:
   a) those variables that are never live during a given range of call sites, and thus need not be considered further for this range.
   b) those variables that are live during the whole range of call sites, and thus can be encoded as a common set for this range of call sites.
   c) those variables whose liveness changes during the range, and which thus need to analyzed in more detail.

For issue 3, it turned out to be pretty simple to change the GC root decoding logic to continue to decode one of the common root strings when it hits one of the encodings described under b). It is just a question of decoding the GC root string index, finding the corresponding GC root string and continuing the decoding there.

[tfs-changeset: 1654187]