Pri-0: Fix custom attribute parsing for arrays within arrays. (#3327)
Scenario:

  class MyAttribute
  {
     public MyAttribute(object[] objs) {}
  }


  [My(new object[] { new int[] { 4, 5, 6} }]
  public void Foo() {...}

Trying to read this with Reflection triggers a
BadImageFormatException as the custom attribute parser was
expecting an array of something, not an Int32Collection.

To understand why this is "Pri-0", take that scenario and do
a few renames...

  class InlineData
  {
     public InlineDataAttribute(params object[] objs) {}
  }

  [Theory]
  [InlineData(new int[] { 4, 5, 6})]
  public static void OneOfThousandsOfCoreFxTestsThatAreFailingOnUwpAot(int[] testdata)
  {
     ...
  }


Unfortunately, there's still one case that's broken
(an element is an enum array - what you get back is
an array of the underlying integral type. I don't
think I can fix this without toolchain work...)

However, even this should help the bring down the
failure count.