Generate stable generic dictionaries in multifile builds
In multifile compilation (when we generate multiple OBJ files with
possibly duplicated content for generics), we need to ensure the generic
dictionary layout generated into each OBJ file is the same, so that
COMDAT folding them at link time causes no harm.

We ensure this by doing two things:

* Always compiling all the methods that contribute to the generic
dictionary. Previously, if code in an OBJ file didn't use a method on a
generic type, we wouldn't compile it. This is not correct for multifile
because the method might be contributing slots to the generic dictionary
of the owning type. We need to make sure it's compiled so that the
dictionary looks the same.
* Sorting the dictionary layout before starting to generate
dictionaries. The order of compilation of methods might not be the same
across different compilations, resulting in different order of slots.
Sorting fixes the problem.

This has no effect on Project X dictionary layout since that one
overrides all the methods that would call into `ComputeLayout`. We
should have really made `DictionaryLayoutNode` an abstract class when
Project X diverged, but that can be cleaned up later. I put it on the
list of my pet peeves (#2645).