Fix warning caused by S.P.CoreLib referencing WindowsRuntimeResourceManager in its .rd.xml.
WindowsRuntimeResourceManager lives in System.Private.WinRTInterop.CoreLib, but it's used from S.P.CoreLib. Since S.P.CoreLib can't have a direct dependency, we use reflection to instantiate it. So we have an .rd.xml directive to force include the WindowsRuntimeResourceManager and make it reflection enabled for activation.
This directive can't line in S.P.CoreLib, since it points to a type outside of its closure. So moving that directive to .rd.xml for S.P.WinRTInterop.CoreLib (and adding the .rd.xml for that assembly since it didn't have one before).
Since we can't have a reference from S.P.CoreLib to S.P.WinRTInterop.CoreLib, we need some other mechanism to force S.P.WinRTInterop.CoreLib to be included by every application (which will mean its .rd.xml is read and thus the WindowsRuntimeResourceManager is "rooted"). This is done by the ILC driver which always includes this dll in the closure when we're building for WinRT.

[tfs-changeset: 1654416]