CoreFx Test Burndown: AssemblyName.FullName Version
Fix AssemblyName.FullName to format assembly names
the way CoreCLR does.

This probably implies we need to review the
binding rules as well but these are sufficiently
non-obvious that I want to treat those separately
(https://github.com/dotnet/corert/issues/3356)
But that's the reason for the separate helper
function visible to AssemblyBinderImplementation.