Port MethodInfo.GetBaseDefinition() quirks.
MethodInfo.GetBaseDefinition() on CoreCLR
produces a MethodHandle internally and
rehydrates it into a MethodInfo only at the end.

(That is, unless the original argument was
declared on an interface type or was non-virtual.)

This had the side effect of removing any generic
method instantation and resetting the ReflectedType
property back to the DeclaredType.

Which the corefx BaseDefinition test was (unwittingly)
relying upon.