Compat fix for ParameterInfo.HasDefaultValue
Parameters without backing metadata (in the
real world, that translates to "return parameters
without custom attributes." but also
parameters and returns of the pseudo-methods
on arrays) report "true" for HasDefaultValue.

This bug has been around in .NET forever
and now CoreFx is testing for that behavior.
It's not worth dying on a hill for.

[tfs-changeset: 1654713]