Treat filter funclets the same as others
Debugging the issues with enabling precise GC info, I found they are
caused by double reporting of locals by both a filter funclet and the
enclosing method. Our code generator for CoreRT treats all funclets
alike, but the runtime expects filter funclets to be treated specially.

This fix lifts the distinction for CoreRT, i.e. treats all funclets the
same. The interaction of GC with EH is pretty complex though, so this is
just a first attempt.