Fix 50 failing System.Runtime.Tests tests.
This had to do with the non-support of InvariantCulture
in various String apis.

I've copy-pasted the InvariantCulture support from
CoreCLR's implementation of the same apis.

Preemptive CR note: I'm intentionally not touching
the other String.Comparison cases in these apis.
Chances are, the CoreCLR code will work fine
(just delegating the job to CultureInfo the same way.)

But I don't own or have any expertise in this code
and I don't want to underwrite the risk of changing
existing code paths or the dead-code cleanup
that's likely to follow.

The only reason I'm doing this one is that this
this is a quick, low-risk change to get a big bump
in our CoreFx-on-UAP pass rates.