Fix stringbuilder marshalling (#3398)
A null parameter for StringBuilder (which is possible in various win32
APIs) argument should skip marshalling. Currently,
`GetEmptyStringBuilderBuffer` assumes an instance, as does the reverse
marshalling code after the call. Adjust `GetEmptyStringBuilderBuffer` to
handle a null StringBuilder, and the native to managed marshaller to
skip replacing the StringBuilder's buffer after the p/invoke.