Switch to generating canonical invoke stubs
This is mostly to address the assert in
`InstantiateDynamicInvokeMethodForMethod` - we shouldn't have a
requirement on whether the method we're generating the stub for is
canonical or not. All the consumers convert the method to the
instantiated form anyway, so we might as well just do it here and avoid
allocating an unnecessary `InstantiatedMethod`.