Fix GetPInvokeDelegateForStub for mcg on CoreCLR
With my recent change to unify delegate marshalling infrastructure between ProjectN and CoreRT, I have made GetPInvokeDelegateForStub in CoreCLR to be a nop. Tijoy pointed out that the part that doesn't need thunk support should work in CoreCLR as it used to be. This trivial change just brought that part for CoreCLR.

[tfs-changeset: 1655953]