Fix hang in Mutex ctor. Reduce requested access rights for Mutex, Semaphore, and EventWaitHandle.
CR: KouVel, SergeyK
Bug: https://devdiv.visualstudio.com/DevDiv/_workitems/edit/406673

Recent IsolatedStorage changes hit an old bug in the Mutex class.  Before .NET Framework 2.0 the constructors of the Mutex class used to simply call the CreateMutex Win32 function.  If the kernel mutex with the given name already exists, CreateMutex requests full access rights to it (MUTEX_ALL_ACCESS).  That request fails if full access rights cannot be granted due to security settings on the mutex.  For instance, creation of the second Mutex below would fail in pre-RTM builds of .NET Framework 2.0:

    var everyoneSid = new SecurityIdentifier(WellKnownSidType.WorldSid, null);
    bool new1, new2;

    // Grant Modify and Synchronize rights to Everyone
    var mtxSec = new MutexSecurity();
    mtxSec.AddAccessRule(new MutexAccessRule(everyoneSid, MutexRights.Modify | MutexRights.Synchronize, AccessControlType.Allow));

    var mtx1 = new Mutex(false, "MyMutex", out new1, mtxSec);
    // Exception: Access is denied.
    var mtx2 = new Mutex(false, "MyMutex", out new2, mtxSec);

In February 2005 there was a change to enable the Mutex constructors to open named mutexes with reduced access rights that allow acquiring and releasing the mutex, but not reading or modifying its security settings:

    VSWhidbey 438936 - Mutex needs to have a loop with CreateMutex/OpenMutex in order to "open or create" a mutex with ACLs
    http://bugcheck/bugs/VSWhidbey/438936
    SD change 1053899; SDPort=DDRTSD:4000

At that pre-Vista time there was no Win32 function for an atomic “create or open the named mutex with the given access rights” operation, so the developer introduced a potentially infinite loop with CreateMutex and OpenMutex calls in it.  The fix assumed that CreateMutex returns ERROR_ACCESS_DENIED only in the case when the named mutex exists, but the caller cannot be granted full access rights to it.  The loop was added for a race condition — the mutex may be deleted between the CreateMutex and OpenMutex calls.

However, CreateMutex also returns ERROR_ACCESS_DENIED if the caller has no permission to create the mutex with the given name.  In that case the changed code calls OpenMutex, which fails to open the mutex because it does not exist, and enters the infinite loop assuming it is constantly losing the race.

A trivial repro for this hang, which demonstrates the bug in .NET Framework and .NET Core as well, is “new Mutex(false, @"Session\foo");” executed in a non-elevated process.  Another repro is “new Mutex(false, @"Global\foo");” executed in an AppContainer.  There are multiple reports of this bug in blogs and forums, e.g., http://stackoverflow.com/questions/36591765/mutex-creation-hangs-while-using-impersonation.

Another unfortunate consequence of the fix above is introduced behavior disparity between the Mutex class on one side and the Semaphore and EventWaitHandle classes on another side.  The scenario fixed above still fails for Semaphore and EventWaitHandle in .NET Framework and .NET Core:

    var everyoneSid = new SecurityIdentifier(WellKnownSidType.WorldSid, null);
    bool new1, new2;

    // Grant Modify and Synchronize rights to Everyone
    var semSec = new SemaphoreSecurity();
    semSec.AddAccessRule(new SemaphoreAccessRule(everyoneSid, SemaphoreRights.Modify | SemaphoreRights.Synchronize, AccessControlType.Allow));

    var sem1 = new Semaphore(0, 1, "MySemaphore", out new1, semSec);
    // UnauthorizedAccessException: Access to the port is denied.
    var sem2 = new Semaphore(0, 1, "MySemaphore", out new2, semSec);

    // Grant Modify and Synchronize rights to Everyone
    var ewhSec = new EventWaitHandleSecurity();
    ewhSec.AddAccessRule(new EventWaitHandleAccessRule(everyoneSid, EventWaitHandleRights.Modify | EventWaitHandleRights.Synchronize, AccessControlType.Allow));

    var ewh1 = new EventWaitHandle(false, EventResetMode.ManualReset, "MyEvent", out new1, ewhSec);
    // UnauthorizedAccessException: Access to the path 'MyEvent' is denied.
    var ewh2 = new EventWaitHandle(false, EventResetMode.ManualReset, "MyEvent", out new2, ewhSec);

In 2006 Windows Vista introduced CreateMutexEx, CreateSemaphoreEx, and CreateEventEx functions implementing previously missing atomic “create or open with the given access rights” operations on named kernel synchronization objects.

In June 2012 TFS change 689674 (http://vstfdevdiv:8080/DevDiv2/DevDiv/_versionControl/changeset/689674) replaced the CreateSemaphore(...) call in the Semaphore constructors with CreateSemaphoreEx(..., SEMAPHORE_MODIFY_STATE | SYNCHRONIZE) — but only in the .NET Native/CoreRT version of the code.  That introduced disparity between .NET Native/CoreRT and .NET Framework/.NET Core.

Note that synchronization operations on Mutex, Semaphore, and EventWaitHandle objects require only (*_MODIFY_STATE | SYNCHRONIZE) access rights.  In fact, [Try]OpenExisting methods of those classes always requested only this minimum set of access rights.  (Strictly speaking, mutexes require only the SYNCHRONIZE access right, and MUTEX_MODIFY_STATE is documented in MSDN as “reserved for future use”.)

However, the .NET Framework also allows reading and modifying security settings of those objects, which requires extra access rights — (READ_CONTROL | WRITE_DAC).  Changing constructors to request only (*_MODIFY_STATE | SYNCHRONIZE) access rights would break that functionality.

On the other hand, .NET Core and CoreRT do not support reading or modifying security settings of kernel synchronization objects, therefore, it makes sense to unify all the constructors and [Try]OpenExisting methods to always request only the minimum set of access rights, (*_MODIFY_STATE | SYNCHRONIZE).  I expect .NET Core to integrate this fix as well.

[tfs-changeset: 1656573]